refactor existing code to use ts (WIP)
