+++ package.json
@@ -2,10 +2,9 @@
     "name": "y",
     "version": "1.0.0",
     "description": "Bitshala discord bot",
-    "main": "src/index.js",
     "scripts": {
         "build": "tsc",
-        "dev": "nodemon src/index.js",
+        "dev": "nodemon dist/src/index.js",
         "lint": "eslint .",
         "lint:fix": "eslint --fix .",
         "format": "prettier --write \"**/*.+(js|jsx|json|css|md)\"",
