+++ src/index.ts
@@ -1,10 +1,10 @@
-const { Client } = require("discord.js");
-const { CommandKit } = require("commandkit");
-const { default: mongoose } = require("mongoose");
+import { Client } from "discord.js";
+import { CommandKit } from "commandkit";
+import { default as mongoose } from "mongoose";
 require("dotenv/config");
 
 const client = new Client({
-    intents: ["Guilds", "GuildMembers", "GuildMessages", "MessageContent"],
+    intents: ["Guilds", "GuildMembers", "GuildMessages", "MessageContent",],
 });
 
 new CommandKit({
@@ -16,7 +16,14 @@ new CommandKit({
     bulkRegister: true,
 });
 
-mongoose.connect(process.env.MONGODB_URI).then(() => {
+const MONGODB_URI = process.env.MONGODB_URI;
+const APP_TOKEN = process.env.APP_TOKEN;
+
+if (!MONGODB_URI || !APP_TOKEN) {
+    throw new Error('Missing environment variables')
+}
+
+mongoose.connect(MONGODB_URI).then(() => {
     console.log("Connected to database.");
-    client.login(process.env.TOKEN);
+    client.login(APP_TOKEN);
 });
